feat: 재사용할 css 변수 및 스타일 컴포넌트 생성
