+++ src/css/mixins.js
@@ -1,0 +1,15 @@
+import { css } from "styled-components";
+
+const flexCenter = css`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
+
+const flexBetween = css`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+`;
+
+export { flexCenter, flexBetween };
