--- src/css/variable.js
@@ -1,21 +1,0 @@
-const color = {
-  Black: "#1b1b1b",
-  GreyOne: "#3f3f3f",
-  GreyTwo: "#777777",
-  GreyThree: "#bcbcbc",
-  GreyFour: "#ebebeb",
-  Green: "#6dd082",
-  Orange: "#ff8e14",
-  OffWhite: "#f8f7f7",
-  White: "#fff",
-};
-
-const fontSize = {
-  xl: "1.5rem",
-  l: "1.25rem",
-  m: "1rem",
-  s: "0.875rem",
-  xs: "0.75rem",
-};
-
-export { color, fontSize };
