+++ src/css/variables.js
@@ -1,0 +1,21 @@
+const color = {
+  black: "#1b1b1b",
+  greyOne: "#3f3f3f",
+  greyTwo: "#777777",
+  greyThree: "#bcbcbc",
+  greyFour: "#ebebeb",
+  green: "#6dd082",
+  orange: "#ff8e14",
+  offWhite: "#f8f7f7",
+  white: "#fff",
+};
+
+const fontSize = {
+  xl: "1.5rem",
+  l: "1.25rem",
+  m: "1rem",
+  s: "0.875rem",
+  xs: "0.75rem",
+};
+
+export { color, fontSize };
